Move useAsync hook
