+++ apps/client-react/src/hooks/use-async.ts
@@ -1,0 +1,110 @@
+import * as React from "react";
+
+export function useSafeDispatch<TAction>(dispatch: React.Dispatch<TAction>) {
+	const mountedRef = React.useRef(false);
+
+	React.useLayoutEffect(() => {
+		mountedRef.current = true;
+		return () => {
+			mountedRef.current = false;
+		};
+	}, []);
+
+	return React.useCallback(
+		(action: TAction) => (mountedRef.current ? dispatch(action) : void 0),
+		[dispatch]
+	);
+}
+
+// Ref: useAsync
+// https://github.com/kentcdodds/bookshelf/blob/main/src/utils/hooks.js
+// https://github.com/kentcdodds/bookshelf/issues/127
+// https://kentcdodds.com/blog/how-to-use-react-context-effectively
+// https://headbutt.io/picking-apart-kents-useasync/
+// https://tanstack.com/query/latest
+export type AsyncState<TData = unknown> =
+	| { status: "PENDING"; data: null; error: null }
+	| { status: "FULFILLED"; data: TData; error: null }
+	| { status: "ERROR"; data: null; error: unknown };
+export type AsyncStatus = AsyncState["status"];
+export type AsyncAction<TData> =
+	| { type: "PENDING" }
+	| { type: "FULFILLED"; data: TData }
+	| { type: "ERROR"; error: unknown }
+	| { type: "RESET"; initialState?: AsyncState<TData> };
+const asyncInitialState: AsyncState = { status: "FULFILLED", data: null, error: null };
+
+export function asyncReducer<TData>(
+	state: AsyncState<TData>,
+	action: AsyncAction<TData>
+): AsyncState<TData> {
+	switch (action.type) {
+		case "PENDING":
+			return { status: "PENDING", data: null, error: null };
+		case "FULFILLED":
+			return { status: "FULFILLED", data: action.data, error: null };
+		case "ERROR":
+			return { status: "ERROR", data: null, error: action.error };
+		case "RESET":
+			return action?.initialState ?? (asyncInitialState as AsyncState<TData>);
+		default:
+			return state;
+	}
+}
+
+export function useAsync<TData>(initialState?: AsyncState<TData>) {
+	const initialStateRef = React.useRef<AsyncState<TData>>({
+		...(asyncInitialState as AsyncState<TData>),
+		...initialState,
+	});
+
+	const [state, unsafeDispatch] = React.useReducer(
+		asyncReducer<TData>,
+		initialStateRef.current
+	);
+	const safeDispatch = useSafeDispatch(unsafeDispatch);
+
+	const setData = React.useCallback(
+		(data: TData) => safeDispatch({ type: "FULFILLED", data }),
+		[safeDispatch]
+	);
+	const setError = React.useCallback(
+		(error: unknown) => safeDispatch({ type: "ERROR", error }),
+		[safeDispatch]
+	);
+	const reset = React.useCallback(
+		() => safeDispatch({ type: "RESET", initialState: initialStateRef.current }),
+		[safeDispatch]
+	);
+
+	const run = React.useCallback(
+		async (promise: Promise<TData>) => {
+			if (!promise || !promise.then)
+				throw new Error("The argument passed to useAsync().run must be a promise.");
+
+			safeDispatch({ type: "PENDING" });
+			try {
+				const data = await promise;
+				setData(data);
+				return data;
+			} catch (error) {
+				setError(error);
+				return await Promise.reject(error);
+			}
+		},
+		[safeDispatch, setData, setError]
+	);
+
+	return {
+		status: state.status,
+		data: state.data,
+		error: state.error,
+		setData,
+		setError,
+		reset,
+		run,
+		isPending: state.status === "PENDING",
+		isFulfilled: state.status === "FULFILLED",
+		isError: state.status === "ERROR",
+	};
+}
